refactor: separate KongState out
